requirements.txt
Pillow==10.4.0
ImageHash==4.3.1
tqdm==4.66.4
PyYAML==6.0.2